move TestExtForcing to test_ext_forcing.py
